app/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "strands-agents[ollama]>=1.8.0",
    "strands-agents-builder>=0.1.9",
    "strands-agents-tools>=0.2.7",
    "fastapi>=0.116.2",
    "uvicorn>=0.35.0",
    "pyinstaller>=6.0.0",
]
